dm-is-state_machine - generic transition! method
This patch adds a transition! method to DataMapper::Is::State::Machine::InstanceMethods that is then called from the dynamically created transition
methods (e.g. forward! in the stoplight example).

This has two benefits. 1) It lets me add before and after hooks to the transition! method that will fire every time a transition occurs and 2) it
allows me to overwrite the forward! method and still call the transition change with transition!